requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
sentence-transformers==3.2.1
numpy==1.26.4
discord.py>=2.5.2
httpx==0.27.2
openai==1.54.4
structlog==24.4.0
rank_bm25==0.2.2
